requirements.txt
Flask==2.2.5
requests==2.31.0
configparser==5.0.2
gunicorn==22.0.0
python-dotenv==1.0.0